+++ src/helpers/block_util.rs
@@ -1,10 +1,9 @@
 use crate::block::BlockContext;
-use crate::error::RenderError;
 use crate::json::value::PathAndJson;
 
 pub(crate) fn create_block<'reg: 'rc, 'rc>(
     param: &'rc PathAndJson<'reg, 'rc>,
-) -> Result<BlockContext<'reg>, RenderError> {
+) -> BlockContext<'reg> {
     let mut block = BlockContext::new();
 
     if let Some(new_path) = param.context_path() {
@@ -14,5 +13,5 @@ pub(crate) fn create_block<'reg: 'rc, 'rc>(
         block.set_base_value(param.value().clone());
     }
 
-    Ok(block)
+    block
 }