+++ src/registry.rs
@@ -1,9 +1,9 @@
+use std::borrow::Cow;
 use std::collections::HashMap;
 use std::fmt::{self, Debug, Formatter};
-use std::fs::File;
-use std::io::prelude::*;
-use std::io::BufReader;
+use std::io::{Error as IoError, Write};
 use std::path::Path;
+use std::sync::Arc;
 
 use serde::Serialize;
 
@@ -11,10 +11,11 @@ use crate::context::Context;
 use crate::decorators::{self, DecoratorDef};
 #[cfg(feature = "script_helper")]
 use crate::error::ScriptError;
-use crate::error::{RenderError, TemplateError, TemplateFileError, TemplateRenderError};
+use crate::error::{RenderError, TemplateError};
 use crate::helpers::{self, HelperDef};
 use crate::output::{Output, StringOutput, WriteOutput};
 use crate::render::{RenderContext, Renderable};
+use crate::sources::{FileSource, Source};
 use crate::support::str::{self, StringWriter};
 use crate::template::Template;
 
@@ -34,7 +35,7 @@ use crate::helpers::scripting::ScriptHelper;
 ///
 /// An *escape fn* is represented as a `Box` to avoid unnecessary type
 /// parameters (and because traits cannot be aliased using `type`).
-pub type EscapeFn = Box<dyn Fn(&str) -> String + Send + Sync>;
+pub type EscapeFn = Arc<dyn Fn(&str) -> String + Send + Sync>;
 
 /// The default *escape fn* replaces the characters `&"<>`
 /// with the equivalent html / xml entities.
@@ -51,15 +52,24 @@ pub fn no_escape(data: &str) -> String {
 /// The single entry point of your Handlebars templates
 ///
 /// It maintains compiled templates and registered helpers.
+#[derive(Clone)]
 pub struct Registry<'reg> {
     templates: HashMap<String, Template>,
-    helpers: HashMap<String, Box<dyn HelperDef + Send + Sync + 'reg>>,
-    decorators: HashMap<String, Box<dyn DecoratorDef + Send + Sync + 'reg>>,
+
+    helpers: HashMap<String, Arc<dyn HelperDef + Send + Sync + 'reg>>,
+    decorators: HashMap<String, Arc<dyn DecoratorDef + Send + Sync + 'reg>>,
+
     escape_fn: EscapeFn,
-    source_map: bool,
     strict_mode: bool,
+    dev_mode: bool,
     #[cfg(feature = "script_helper")]
-    pub(crate) engine: Engine,
+    pub(crate) engine: Arc<Engine>,
+
+    template_sources:
+        HashMap<String, Arc<dyn Source<Item = String, Error = IoError> + Send + Sync + 'reg>>,
+    #[cfg(feature = "script_helper")]
+    script_sources:
+        HashMap<String, Arc<dyn Source<Item = String, Error = IoError> + Send + Sync + 'reg>>,
 }
 
 impl<'reg> Debug for Registry<'reg> {
@@ -68,7 +78,8 @@ impl<'reg> Debug for Registry<'reg> {
             .field("templates", &self.templates)
             .field("helpers", &self.helpers.keys())
             .field("decorators", &self.decorators.keys())
-            .field("source_map", &self.source_map)
+            .field("strict_mode", &self.strict_mode)
+            .field("dev_mode", &self.dev_mode)
             .finish()
     }
 }
@@ -103,13 +114,16 @@ impl<'reg> Registry<'reg> {
     pub fn new() -> Registry<'reg> {
         let r = Registry {
             templates: HashMap::new(),
+            template_sources: HashMap::new(),
             helpers: HashMap::new(),
             decorators: HashMap::new(),
-            escape_fn: Box::new(html_escape),
-            source_map: true,
+            escape_fn: Arc::new(html_escape),
             strict_mode: false,
+            dev_mode: false,
             #[cfg(feature = "script_helper")]
-            engine: rhai_engine(),
+            engine: Arc::new(rhai_engine()),
+            #[cfg(feature = "script_helper")]
+            script_sources: HashMap::new(),
         };
 
         r.setup_builtins()
@@ -138,25 +152,15 @@ impl<'reg> Registry<'reg> {
         self
     }
 
-    /// Enable handlebars template source map
-    ///
-    /// Source map provides line/col reporting on error. It uses slightly
-    /// more memory to maintain the data.
-    ///
-    /// Default is true.
-    pub fn source_map_enabled(&mut self, enable: bool) {
-        self.source_map = enable;
-    }
-
-    /// Enable handlebars strict mode
+    /// Enable or disable handlebars strict mode
     ///
     /// By default, handlebars renders empty string for value that
     /// undefined or never exists. Since rust is a static type
     /// language, we offer strict mode in handlebars-rust.  In strict
     /// mode, if you were to render a value that doesn't exist, a
     /// `RenderError` will be raised.
-    pub fn set_strict_mode(&mut self, enable: bool) {
-        self.strict_mode = enable;
+    pub fn set_strict_mode(&mut self, enabled: bool) {
+        self.strict_mode = enabled;
     }
 
     /// Return strict mode state, default is false.
@@ -168,6 +172,22 @@ impl<'reg> Registry<'reg> {
     /// `RenderError` will be raised.
     pub fn strict_mode(&self) -> bool {
         self.strict_mode
+    }
+
+    /// Return dev mode state, default is false
+    ///
+    /// With dev mode turned on, handlebars enables a set of development
+    /// firendly features, that may affect its performance.
+    pub fn dev_mode(&self) -> bool {
+        self.dev_mode
+    }
+
+    /// Enable or disable dev mode
+    ///
+    /// With dev mode turned on, handlebars enables a set of development
+    /// firendly features, that may affect its performance.
+    pub fn set_dev_mode(&mut self, enabled: bool) {
+        self.dev_mode = enabled;
     }
 
     /// Register a `Template`
@@ -190,7 +210,7 @@ impl<'reg> Registry<'reg> {
     where
         S: AsRef<str>,
     {
-        let template = Template::compile_with_name(tpl_str, name.to_owned(), self.source_map)?;
+        let template = Template::compile_with_name(tpl_str, name.to_owned())?;
         self.register_template(name, template);
         Ok(())
     }
@@ -211,14 +231,22 @@ impl<'reg> Registry<'reg> {
         &mut self,
         name: &str,
         tpl_path: P,
-    ) -> Result<(), TemplateFileError>
+    ) -> Result<(), TemplateError>
     where
         P: AsRef<Path>,
     {
-        let mut reader = BufReader::new(
-            File::open(tpl_path).map_err(|e| TemplateFileError::IOError(e, name.to_owned()))?,
-        );
-        self.register_template_source(name, &mut reader)
+        let source = FileSource::new(tpl_path.as_ref().into());
+        let template_string = source
+            .load()
+            .map_err(|err| TemplateError::from((err, name.to_owned())))?;
+
+        self.register_template_string(name, template_string)?;
+        if self.dev_mode {
+            self.template_sources
+                .insert(name.to_owned(), Arc::new(source));
+        }
+
+        Ok(())
     }
 
     /// Register templates from a directory
@@ -238,7 +266,7 @@ impl<'reg> Registry<'reg> {
         &mut self,
         tpl_extension: &'static str,
         dir_path: P,
-    ) -> Result<(), TemplateFileError>
+    ) -> Result<(), TemplateError>
     where
         P: AsRef<Path>,
     {
@@ -271,35 +299,14 @@ impl<'reg> Registry<'reg> {
         Ok(())
     }
 
-    /// Register a template from `std::io::Read` source
-    pub fn register_template_source<R>(
-        &mut self,
-        name: &str,
-        tpl_source: &mut R,
-    ) -> Result<(), TemplateFileError>
-    where
-        R: Read,
-    {
-        let mut buf = String::new();
-        tpl_source
-            .read_to_string(&mut buf)
-            .map_err(|e| TemplateFileError::IOError(e, name.to_owned()))?;
-        self.register_template_string(name, buf)?;
-        Ok(())
-    }
-
     /// Remove a template from the registry
     pub fn unregister_template(&mut self, name: &str) {
         self.templates.remove(name);
     }
 
     /// Register a helper
-    pub fn register_helper(
-        &mut self,
-        name: &str,
-        def: Box<dyn HelperDef + Send + Sync + 'reg>,
-    ) -> Option<Box<dyn HelperDef + Send + Sync + 'reg>> {
-        self.helpers.insert(name.to_string(), def)
+    pub fn register_helper(&mut self, name: &str, def: Box<dyn HelperDef + Send + Sync + 'reg>) {
+        self.helpers.insert(name.to_string(), def.into());
     }
 
     /// Register a [rhai](https://docs.rs/rhai/) script as handlebars helper
@@ -324,16 +331,12 @@ impl<'reg> Registry<'reg> {
     ///
     ///
     #[cfg(feature = "script_helper")]
-    pub fn register_script_helper(
-        &mut self,
-        name: &str,
-        script: String,
-    ) -> Result<Option<Box<dyn HelperDef + Send + Sync + 'reg>>, ScriptError> {
-        let compiled = self.engine.compile(&script)?;
+    pub fn register_script_helper(&mut self, name: &str, script: &str) -> Result<(), ScriptError> {
+        let compiled = self.engine.compile(script)?;
         let script_helper = ScriptHelper { script: compiled };
-        Ok(self
-            .helpers
-            .insert(name.to_string(), Box::new(script_helper)))
+        self.helpers
+            .insert(name.to_string(), Arc::new(script_helper));
+        Ok(())
     }
 
     /// Register a [rhai](https://docs.rs/rhai/) script from file
@@ -342,17 +345,16 @@ impl<'reg> Registry<'reg> {
         &mut self,
         name: &str,
         script_path: P,
-    ) -> Result<Option<Box<dyn HelperDef + Send + Sync + 'reg>>, ScriptError>
+    ) -> Result<(), ScriptError>
     where
         P: AsRef<Path>,
     {
-        let mut script = String::new();
-        {
-            let mut file = File::open(script_path)?;
-            file.read_to_string(&mut script)?;
-        }
+        let source = FileSource::new(script_path.as_ref().into());
+        let script = source.load()?;
 
-        self.register_script_helper(name, script)
+        self.script_sources
+            .insert(name.to_owned(), Arc::new(source));
+        self.register_script_helper(name, &script)
     }
 
     /// Register a decorator
@@ -360,8 +362,8 @@ impl<'reg> Registry<'reg> {
         &mut self,
         name: &str,
         def: Box<dyn DecoratorDef + Send + Sync + 'reg>,
-    ) -> Option<Box<dyn DecoratorDef + Send + Sync + 'reg>> {
-        self.decorators.insert(name.to_string(), def)
+    ) {
+        self.decorators.insert(name.to_string(), def.into());
     }
 
     /// Register a new *escape fn* to be used from now on by this registry.
@@ -369,12 +371,12 @@ impl<'reg> Registry<'reg> {
         &mut self,
         escape_fn: F,
     ) {
-        self.escape_fn = Box::new(escape_fn);
+        self.escape_fn = Arc::new(escape_fn);
     }
 
     /// Restore the default *escape fn*.
     pub fn unregister_escape_fn(&mut self) {
-        self.escape_fn = Box::new(html_escape);
+        self.escape_fn = Arc::new(html_escape);
     }
 
     /// Get a reference to the current *escape fn*.
@@ -392,9 +394,43 @@ impl<'reg> Registry<'reg> {
         self.templates.get(name)
     }
 
+    #[inline]
+    fn get_or_load_template(&'reg self, name: &str) -> Result<Cow<'reg, Template>, RenderError> {
+        if let (true, Some(source)) = (self.dev_mode, self.template_sources.get(name)) {
+            source
+                .load()
+                .map_err(|e| TemplateError::from((e, name.to_owned())))
+                .and_then(|tpl_str| Template::compile_with_name(tpl_str, name.to_owned()))
+                .map(Cow::Owned)
+                .map_err(RenderError::from)
+        } else {
+            self.templates
+                .get(name)
+                .map(Cow::Borrowed)
+                .ok_or_else(|| RenderError::new(format!("Template not found: {}", name)))
+        }
+    }
+
     /// Return a registered helper
-    pub fn get_helper(&self, name: &str) -> Option<&(dyn HelperDef + Send + Sync + 'reg)> {
-        self.helpers.get(name).map(|v| v.as_ref())
+    pub(crate) fn get_or_load_helper(
+        &'reg self,
+        name: &str,
+    ) -> Result<Option<Arc<dyn HelperDef + Send + Sync + 'reg>>, RenderError> {
+        #[cfg(feature = "script_helper")]
+        if let (true, Some(source)) = (self.dev_mode, self.script_sources.get(name)) {
+            return source
+                .load()
+                .map_err(ScriptError::from)
+                .and_then(|s| {
+                    let helper = Box::new(ScriptHelper {
+                        script: self.engine.compile(&s)?,
+                    }) as Box<dyn HelperDef + Send + Sync>;
+                    Ok(Some(helper.into()))
+                })
+                .map_err(RenderError::from);
+        }
+
+        Ok(self.helpers.get(name).cloned())
     }
 
     #[inline]
@@ -403,7 +439,10 @@ impl<'reg> Registry<'reg> {
     }
 
     /// Return a registered decorator
-    pub fn get_decorator(&self, name: &str) -> Option<&(dyn DecoratorDef + Send + Sync + 'reg)> {
+    pub(crate) fn get_decorator(
+        &self,
+        name: &str,
+    ) -> Option<&(dyn DecoratorDef + Send + Sync + 'reg)> {
         self.decorators.get(name).map(|v| v.as_ref())
     }
 
@@ -417,6 +456,7 @@ impl<'reg> Registry<'reg> {
         self.templates.clear();
     }
 
+    #[inline]
     fn render_to_output<O>(
         &self,
         name: &str,
@@ -426,12 +466,10 @@ impl<'reg> Registry<'reg> {
     where
         O: Output,
     {
-        self.get_template(name)
-            .ok_or_else(|| RenderError::new(format!("Template not found: {}", name)))
-            .and_then(|t| {
-                let mut render_context = RenderContext::new(t.name.as_ref());
-                t.render(self, &ctx, &mut render_context, output)
-            })
+        self.get_or_load_template(name).and_then(|t| {
+            let mut render_context = RenderContext::new(t.name.as_ref());
+            t.render(self, &ctx, &mut render_context, output)
+        })
     }
 
     /// Render a registered template with some data into a string
@@ -469,11 +507,7 @@ impl<'reg> Registry<'reg> {
     }
 
     /// Render a template string using current registry without registering it
-    pub fn render_template<T>(
-        &self,
-        template_string: &str,
-        data: &T,
-    ) -> Result<String, TemplateRenderError>
+    pub fn render_template<T>(&self, template_string: &str, data: &T) -> Result<String, RenderError>
     where
         T: Serialize,
     {
@@ -487,8 +521,8 @@ impl<'reg> Registry<'reg> {
         &self,
         template_string: &str,
         ctx: &Context,
-    ) -> Result<String, TemplateRenderError> {
-        let tpl = Template::compile2(template_string, self.source_map)?;
+    ) -> Result<String, RenderError> {
+        let tpl = Template::compile(template_string)?;
 
         let mut out = StringOutput::new();
         {
@@ -496,8 +530,7 @@ impl<'reg> Registry<'reg> {
             tpl.render(self, &ctx, &mut render_context, &mut out)?;
         }
 
-        out.into_string()
-            .map_err(|e| TemplateRenderError::from(RenderError::from(e)))
+        out.into_string().map_err(RenderError::from)
     }
 
     /// Render a template string using current registry without registering it
@@ -506,36 +539,16 @@ impl<'reg> Registry<'reg> {
         template_string: &str,
         data: &T,
         writer: W,
-    ) -> Result<(), TemplateRenderError>
+    ) -> Result<(), RenderError>
     where
         T: Serialize,
         W: Write,
     {
-        let tpl = Template::compile2(template_string, self.source_map)?;
+        let tpl = Template::compile(template_string)?;
         let ctx = Context::wraps(data)?;
         let mut render_context = RenderContext::new(None);
         let mut out = WriteOutput::new(writer);
         tpl.render(self, &ctx, &mut render_context, &mut out)
-            .map_err(TemplateRenderError::from)
-    }
-
-    /// Render a template source using current registry without registering it
-    pub fn render_template_source_to_write<T, R, W>(
-        &self,
-        template_source: &mut R,
-        data: &T,
-        writer: W,
-    ) -> Result<(), TemplateRenderError>
-    where
-        T: Serialize,
-        W: Write,
-        R: Read,
-    {
-        let mut tpl_str = String::new();
-        template_source
-            .read_to_string(&mut tpl_str)
-            .map_err(|e| TemplateRenderError::IOError(e, "Unnamed template source".to_owned()))?;
-        self.render_template_to_write(&tpl_str, data, writer)
     }
 }
 
@@ -549,11 +562,8 @@ mod test {
     use crate::render::{Helper, RenderContext, Renderable};
     use crate::support::str::StringWriter;
     use crate::template::Template;
-    #[cfg(feature = "dir_source")]
-    use std::fs::{DirBuilder, File};
-    #[cfg(feature = "dir_source")]
+    use std::fs::File;
     use std::io::Write;
-    #[cfg(feature = "dir_source")]
     use tempfile::tempdir;
 
     #[derive(Clone, Copy)]
@@ -606,6 +616,8 @@ mod test {
     #[test]
     #[cfg(feature = "dir_source")]
     fn test_register_templates_directory() {
+        use std::fs::DirBuilder;
+
         let mut r = Registry::new();
         {
             let dir = tempdir().unwrap();
@@ -782,10 +794,7 @@ mod test {
         let render_error = r
             .render_template("accessing non-exists key {{the_key_never_exists}}", &data)
             .unwrap_err();
-        assert_eq!(
-            render_error.as_render_error().unwrap().column_no.unwrap(),
-            26
-        );
+        assert_eq!(render_error.column_no.unwrap(), 26);
 
         let data2 = json!([1, 2, 3]);
         assert!(r
@@ -797,10 +806,7 @@ mod test {
         let render_error2 = r
             .render_template("accessing invalid array index {{this.[3]}}", &data2)
             .unwrap_err();
-        assert_eq!(
-            render_error2.as_render_error().unwrap().column_no.unwrap(),
-            31
-        );
+        assert_eq!(render_error2.column_no.unwrap(), 31);
     }
 
     use crate::json::value::ScopedJson;
@@ -958,5 +964,85 @@ mod test {
             )
             .unwrap()
         );
+    }
+
+    #[test]
+    fn test_dev_mode_template_reload() {
+        let mut reg = Registry::new();
+        reg.set_dev_mode(true);
+        assert!(reg.dev_mode());
+
+        let dir = tempdir().unwrap();
+        let file1_path = dir.path().join("t1.hbs");
+        {
+            let mut file1: File = File::create(&file1_path).unwrap();
+            write!(file1, "<h1>Hello {{{{name}}}}!</h1>").unwrap();
+        }
+
+        reg.register_template_file("t1", &file1_path).unwrap();
+
+        assert_eq!(
+            reg.render("t1", &json!({"name": "Alex"})).unwrap(),
+            "<h1>Hello Alex!</h1>"
+        );
+
+        {
+            let mut file1: File = File::create(&file1_path).unwrap();
+            write!(file1, "<h1>Privet {{{{name}}}}!</h1>").unwrap();
+        }
+
+        assert_eq!(
+            reg.render("t1", &json!({"name": "Alex"})).unwrap(),
+            "<h1>Privet Alex!</h1>"
+        );
+
+        dir.close().unwrap();
+    }
+
+    #[test]
+    #[cfg(feature = "script_helper")]
+    fn test_script_helper() {
+        let mut reg = Registry::new();
+
+        reg.register_script_helper("acc", "params.reduce(|sum, x| x + sum, || 0 )")
+            .unwrap();
+
+        assert_eq!(
+            reg.render_template("{{acc 1 2 3 4}}", &json!({})).unwrap(),
+            "10"
+        );
+    }
+
+    #[test]
+    #[cfg(feature = "script_helper")]
+    fn test_script_helper_dev_mode() {
+        let mut reg = Registry::new();
+        reg.set_dev_mode(true);
+
+        let dir = tempdir().unwrap();
+        let file1_path = dir.path().join("acc.rhai");
+        {
+            let mut file1: File = File::create(&file1_path).unwrap();
+            write!(file1, "params.reduce(|sum, x| x + sum, || 0 )").unwrap();
+        }
+
+        reg.register_script_helper_file("acc", &file1_path).unwrap();
+
+        assert_eq!(
+            reg.render_template("{{acc 1 2 3 4}}", &json!({})).unwrap(),
+            "10"
+        );
+
+        {
+            let mut file1: File = File::create(&file1_path).unwrap();
+            write!(file1, "params.reduce(|sum, x| x * sum, || 1 )").unwrap();
+        }
+
+        assert_eq!(
+            reg.render_template("{{acc 1 2 3 4}}", &json!({})).unwrap(),
+            "24"
+        );
+
+        dir.close().unwrap();
     }
 }